+++ tests/serialization/test_serialize_series_to_toml.py
@@ -7,7 +7,7 @@ from textwrap import dedent
 
 import pytest
 
-from lanpartydb.models import Series
+from lanpartydb.models import Series, SeriesLinks, Resource
 from lanpartydb.serialization import serialize_series_to_toml
 
 
@@ -33,12 +33,42 @@ from lanpartydb.serialization import serialize_series_to_toml
                 title='DeltaLAN',
                 alternative_titles=['Δ LAN', 'Δέλτα LAN'],
                 country_codes=['au'],
+                links=SeriesLinks(
+                    website=Resource(
+                        url='https://www.deltalan.example/',
+                    ),
+                ),
             ),
             dedent("""\
             slug = "deltalan"
             title = "DeltaLAN"
             alternative_titles = ["Δ LAN", "Δέλτα LAN"]
             country_codes = ["au"]
+
+            [links.website]
+            url = "https://www.deltalan.example/"
+            """),
+        ),
+        (
+            Series(
+                slug='epsilan',
+                title='EpsiLAN',
+                country_codes=['gr'],
+                links=SeriesLinks(
+                    website=Resource(
+                        url='https://www.epsilan.example/',
+                        offline=True,
+                    ),
+                ),
+            ),
+            dedent("""\
+            slug = "epsilan"
+            title = "EpsiLAN"
+            country_codes = ["gr"]
+
+            [links.website]
+            url = "https://www.epsilan.example/"
+            offline = true
             """),
         ),
     ],